some refactor, autocomplete still being a jerk, may need to rethink
