+++ src/components/NavBar.js
@@ -1,12 +1,11 @@
 import React from 'react';
 import {Navbar, NavItem, Icon} from 'react-materialize';
-import { version } from 'variadic.js/package.json';
 
-const NavBar = () => (
+const NavBar = (props) => (
   <Navbar
   	right
     style={{backgroundColor: "#424242", paddingLeft:"20px"}}
-    brand={`VariadicJS (${version})`}
+    brand={props.version}
     >
     <NavItem>
         <Icon medium>settings</Icon>
